支持para打开分布式和低时延dump

+++ frameworks/native/audiostream/include/fast_audio_stream.h
@@ -113,7 +113,7 @@ public:
 
     // Playback related APIs
     bool DrainAudioStream() override;
-    size_t Write(uint8_t *buffer, size_t buffer_size) override;
+    int32_t Write(uint8_t *buffer, size_t buffer_size) override;
 
     // Recording related APIs
     int32_t Read(uint8_t &buffer, size_t userSize, bool isBlockingRead) override;
